+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Apache Shiro authn/authz support for `jax-rs`
 - Apache Shiro authn/authz support for `spring`
 - Passport authn support for `express`
+- `express-openid-connect` authn support for `express`
+- `express-jwt` authn support for `express`
 
 ## [0.5.0] - 2023-03-14